@@ -1,18 +1,21 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.3.71'
-    ext.base_kt_version = "eca8321c52"
+    ext.kotlin_version = '1.4.21'
+    ext.base_kt_version = "1.0.6"
     ext.SimXutilsDb =     "77f2bade32"
+
+    ext.recorder_version = '0.0.4'
+
     repositories {
         google()
         jcenter()
-
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.2'
+        classpath 'com.android.tools.build:gradle:4.1.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1' // Add this line
+        classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.5'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -23,6 +26,9 @@ allprojects {
         google()
         jcenter()
         maven { url "https://jitpack.io" }
+        maven{
+            url 'https://dl.bintray.com/shetiejun/maven'
+        }
     }
 }
 
