@@ -2,46 +2,75 @@ package me.shetj.mp3recorder
 
 import android.Manifest
 import android.os.Bundle
-import android.util.Log
-import androidx.appcompat.app.AppCompatActivity
-import com.jakewharton.rxbinding3.view.clicks
-import com.tbruyelle.rxpermissions2.RxPermissions
-import kotlinx.android.synthetic.main.activity_main.*
-import me.shetj.base.tools.app.ArmsUtils
+import me.shetj.base.ktx.hasPermission
+import me.shetj.base.ktx.showToast
+import me.shetj.base.ktx.start
+import me.shetj.base.mvvm.BaseBindingActivity
+import me.shetj.base.mvvm.BaseViewModel
+import me.shetj.mp3recorder.databinding.ActivityMainBinding
 import me.shetj.mp3recorder.record.MixRecordActivity
 import me.shetj.mp3recorder.record.activity.mix.MyMixRecordActivity
-import me.shetj.recorder.util.LameUtils
+import me.shetj.mp3recorder.record.activity.sim.MyRecordActivity
+import me.shetj.recorder.ui.RecorderPopup
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : BaseBindingActivity<BaseViewModel, ActivityMainBinding>() {
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        Log.i("MainActivity",LameUtils.version())
-        RxPermissions(this).request(
-            Manifest.permission.WRITE_EXTERNAL_STORAGE,Manifest.permission.READ_EXTERNAL_STORAGE
-        ,Manifest.permission.RECORD_AUDIO).subscribe()
+    private val recorderPopup: RecorderPopup by lazy {
+        RecorderPopup(this, needPlay = false, maxTime = (60 * 60 * 1000).toLong()) {
+            it.showToast()
+        }
+    }
 
-        btn_demo.clicks()
-            .compose(RxPermissions(this).ensure(Manifest.permission.WRITE_EXTERNAL_STORAGE,Manifest.permission.READ_EXTERNAL_STORAGE
-                ,Manifest.permission.RECORD_AUDIO))
-            .subscribe {
-                ArmsUtils.startActivity(this,
-                    MyMixRecordActivity::class.java)
+    override fun onActivityCreate() {
+        super.onActivityCreate()
+        hasPermission(
+            Manifest.permission.WRITE_EXTERNAL_STORAGE,
+            Manifest.permission.READ_EXTERNAL_STORAGE,
+            Manifest.permission.RECORD_AUDIO,
+            isRequest = true
+        )
+
+        mViewBinding.btnDemo.setOnClickListener {
+            if (hasPermission(
+                    Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                    Manifest.permission.READ_EXTERNAL_STORAGE,
+                    Manifest.permission.RECORD_AUDIO, isRequest = true
+                )
+            ) {
+                start<MyRecordActivity>()
             }
-
-        btn_demo2.clicks()
-            .compose(RxPermissions(this).ensure(Manifest.permission.WRITE_EXTERNAL_STORAGE,Manifest.permission.READ_EXTERNAL_STORAGE
-                ,Manifest.permission.RECORD_AUDIO))
-            .subscribe {
-            ArmsUtils.startActivity(this,MixRecordActivity::class.java)
         }
 
-        btn_demo3.clicks()
-            .compose(RxPermissions(this).ensure(Manifest.permission.WRITE_EXTERNAL_STORAGE,Manifest.permission.READ_EXTERNAL_STORAGE
-                ,Manifest.permission.RECORD_AUDIO))
-            .subscribe {
-                ArmsUtils.startActivity(this,MyMixRecordActivity::class.java)
+        mViewBinding.btnDemo2.setOnClickListener {
+            if (hasPermission(
+                    Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                    Manifest.permission.READ_EXTERNAL_STORAGE,
+                    Manifest.permission.RECORD_AUDIO, isRequest = true
+                )
+            ) {
+                start<MixRecordActivity>()
             }
+        }
+
+        mViewBinding.btnDemo3.setOnClickListener {
+            if (hasPermission(
+                    Manifest.permission.WRITE_EXTERNAL_STORAGE,
+                    Manifest.permission.READ_EXTERNAL_STORAGE,
+                    Manifest.permission.RECORD_AUDIO, isRequest = true
+                )
+            ) {
+                start<MyMixRecordActivity>()
+            }
+        }
+
+        mViewBinding.btnDemo4.setOnClickListener {
+            recorderPopup.showPop()
+        }
+    }
+
+    override fun onBackPressed() {
+        if (recorderPopup.onBackPress()) {
+            super.onBackPressed()
+        }
     }
 }
