@@ -9,31 +9,32 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.inputmethod.EditorInfo
 import android.widget.*
-import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import com.chad.library.adapter.base.listener.OnItemClickListener
-import io.reactivex.BackpressureStrategy
-import io.reactivex.Flowable
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
+import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
+import io.reactivex.rxjava3.core.BackpressureStrategy
+import io.reactivex.rxjava3.core.Flowable
+import io.reactivex.rxjava3.schedulers.Schedulers
+import me.shetj.base.ktx.logi
 import me.shetj.base.tools.app.ArmsUtils
+import me.shetj.dialog.OrangeDialog
 import me.shetj.mp3recorder.R
-import me.shetj.mp3recorder.record.bean.Music
+import me.shetj.mp3recorder.record.bean.MusicQ
 import me.shetj.mp3recorder.record.bean.Record
 import me.shetj.mp3recorder.record.bean.RecordDbUtils
 import me.shetj.mp3recorder.record.utils.*
 import me.shetj.mp3recorder.record.view.BackgroundMixMusicView
-import me.shetj.mp3recorder.record.view.MusicListBottomSheetDialog
-import org.simple.eventbus.EventBus
-import org.simple.eventbus.Subscriber
-import org.simple.eventbus.ThreadMode
+import me.shetj.mp3recorder.record.view.MusicListBottomQSheetDialog
+import me.shetj.recorder.core.SimRecordListener
+import org.greenrobot.eventbus.EventBus
+import org.greenrobot.eventbus.Subscribe
+import org.greenrobot.eventbus.ThreadMode
 import timber.log.Timber
 import java.io.File
 
 /**
  * 录制声音界面
  */
-open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGroup, private val callback: Callback) : View.OnClickListener {
+open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGroup, private val callback: EventCallback) : View.OnClickListener {
 
     private val root: RelativeLayout = LayoutInflater.from(context).inflate(R.layout.page_record_mix, null) as RelativeLayout
     private var mEditInfo: EditText? = null
@@ -48,16 +49,16 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
     val scene: Scene
     private var mtvAllTime: TextView? = null
     private var mBunceSv: ScrollView? = null
-    private var recordCallBack: RecordCallBack? = null
+    private var recordCallBack: SimRecordListener? = null
     private var musicView: BackgroundMixMusicView?=null
     private var addMusic :LinearLayout ?=null
-    private val musicDialog: MusicListBottomSheetDialog by lazy {
-        MusicListBottomSheetDialog(context).apply {
-            setOnItemClickListener( OnItemClickListener { adapter, _, position ->
-                val music = adapter.getItem(position) as Music
+    private val musicDialog: MusicListBottomQSheetDialog by lazy {
+        MusicListBottomQSheetDialog(context).apply {
+            setOnItemClickListener { adapter, _, position ->
+                val music = adapter.getItem(position) as MusicQ
                 musicView?.setMusic(music)
                 addMusic?.visibility = View.GONE
-            })
+            }
         }
     }
     private var recordUtils:MixRecordUtils?=null
@@ -88,7 +89,7 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
         addMusic!!.setOnClickListener(this)
     }
 
-    @Subscriber(mode = ThreadMode.MAIN)
+    @Subscribe(threadMode= ThreadMode.MAIN)
     fun refreshData(event: MainThreadEvent<*>) {
         when (event.type) {
             MainThreadEvent.REMOVE_MUSIC -> {
@@ -106,11 +107,11 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
         val editText = EditText(this.context)
         val params = ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT)
         editText.layoutParams = params
-        editText.minHeight = ArmsUtils.dip2px(500f)
+        editText.minHeight = ArmsUtils.dp2px(500f)
         editText.textSize = 16f
         editText.gravity = Gravity.TOP
         editText.setBackgroundColor(Color.WHITE)
-        editText.setPadding(ArmsUtils.dip2px(30f), ArmsUtils.dip2px(27f), ArmsUtils.dip2px(30f), ArmsUtils.dip2px(50f))
+        editText.setPadding(ArmsUtils.dp2px(30f), ArmsUtils.dp2px(27f), ArmsUtils.dp2px(30f), ArmsUtils.dp2px(50f))
         editText.inputType = EditorInfo.TYPE_CLASS_TEXT or EditorInfo.TYPE_TEXT_FLAG_MULTI_LINE
         editText.setHorizontallyScrolling(false)
         editText.maxLines = Integer.MAX_VALUE
@@ -126,8 +127,10 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
     }
 
     private fun initData() {
-        recordCallBack = object : RecordCallBack {
-            override fun start() {
+        recordCallBack = object : SimRecordListener() {
+
+            override fun onStart() {
+                super.onStart()
                 TransitionManager.beginDelayedTransition(root)
                 isHasChange = true
                 mIvRecordState!!.setImageResource(R.mipmap.icon_record_pause_2)
@@ -136,34 +139,37 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
 
             }
 
-            /**
-             * 如果要写呼吸灯，就在这里处理
-             * @param time
-             * @param volume
-             */
-            override fun onRecording(time: Int, volume: Int) {
-                Timber.i( "time = $time\nvolume$volume")
+            override fun onRemind(duration: Long) {
+                super.onRemind(duration)
+                "这是onRemind ：${Util.formatSeconds3((duration/1000).toInt())}".logi()
+
+            }
+
+            override fun onRecording(time: Long, volume: Int) {
+                super.onRecording(time, volume)
+//                Timber.i( "time = $time\nvolume$volume")
                 AndroidSchedulers.mainThread().scheduleDirect {
-                    mProgressBarRecord!!.progress = time
-                    mTvRecordTime!!.text = Util.formatSeconds3(time)
+                    mProgressBarRecord!!.progress = time.toInt()
+                    mTvRecordTime!!.text = Util.formatSeconds3(time.toInt())
                 }
             }
 
-            override fun pause() {
+            override fun onPause() {
                 TransitionManager.beginDelayedTransition(root)
                 mIvRecordState!!.setImageResource(R.mipmap.icon_start_record)
                 showSaveAndRe(View.VISIBLE)
                 mTvStateMsg!!.text = "已暂停，点击继续"
             }
 
-            override fun onSuccess(file: String, time: Int) {
+            override fun onSuccess(file: String, time: Long) {
+                Timber.i( "onSuccess")
                 if (File(file).exists()) {
                     TransitionManager.beginDelayedTransition(root)
                     mIvRecordState!!.setImageResource(R.mipmap.icon_start_record)
                     showSaveAndRe(View.INVISIBLE)
                     mTvStateMsg!!.text = "点击录音"
                     if (oldRecord == null) {
-                        saveRecord(file, time)
+                        saveRecord(file)
                         callback.onEvent(1)
                     } else {
                         saveOldRecord(file, true)
@@ -171,14 +177,9 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
                 }
             }
 
-            override fun onProgress(time: Int) {
-                mProgressBarRecord!!.progress = time
-                mTvRecordTime!!.text = Util.formatSeconds3(time)
-            }
-
-            override fun onMaxProgress(time: Int) {
-                mProgressBarRecord!!.max = time
-                mtvAllTime!!.text = Util.formatSeconds3(time)
+            override fun onMaxChange(time: Long) {
+                mProgressBarRecord!!.max = time.toInt()
+                mtvAllTime!!.text = Util.formatSeconds3(time.toInt())
             }
 
             override fun onError(e: Exception) {
@@ -186,14 +187,14 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
                 setRecord(oldRecord)
             }
 
-            override fun autoComplete(file: String, time: Int) {
+            override fun autoComplete(file: String, time: Long) {
                 if (File(file).exists()) {
                     TransitionManager.beginDelayedTransition(root)
                     mIvRecordState!!.setImageResource(R.mipmap.icon_start_record)
                     showSaveAndRe(View.INVISIBLE)
                     mTvStateMsg!!.text = "录制完成"
                     if (oldRecord == null) {
-                        saveRecord(file, time)
+                        saveRecord(file)
                         showRecordNewDialog()
                     } else {
                         saveOldRecord(file, false)
@@ -242,11 +243,13 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
      * 是否录制新内容
      */
     private fun showRecordNewDialog() {
-        AlertDialog.Builder(context)
+        OrangeDialog.Builder(context)
             .setTitle("录音已保存")
-            .setMessage("已成功录满${1200/60}分钟，录音已保存。是否继续录制下一条？")
-            .setNegativeButton("查看本条") { _, _ -> callback.onEvent(1) }
-            .setPositiveButton("录下一条") { _, _ ->
+            .setContent("已成功录满${1200/60}分钟，录音已保存。是否继续录制下一条？")
+            .setNegativeText("查看本条")
+            .setOnNegativeCallBack {  _, _ -> callback.onEvent(1) }
+            .setPositiveText("录下一条")
+            .setonPositiveCallBack {  _, _ ->
                 setRecord(null)
                 ArmsUtils.makeText("上条录音已保存至“我的录音”")
             }
@@ -257,7 +260,7 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
     //得到media的时间长度
     private fun getMediaTime(s: String): Flowable<Int> {
         return Flowable.create({ emitter ->
-            val audioLength = Util.getAudioLength(context, s)
+            val audioLength = Util.getAudioLength(s)
             emitter.onNext(audioLength)
         }, BackpressureStrategy.BUFFER)
     }
@@ -266,10 +269,10 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
     /**
      * 保存录音，并且通知修改
      */
-    private fun saveRecord(file: String, time: Int) {
+    private fun saveRecord(file: String) {
         try {
             val record = Record("1", file, System.currentTimeMillis().toString() + "",
-                Util.getAudioLength(context, file), mEditInfo!!.text.toString())
+                Util.getAudioLength(file), mEditInfo!!.text.toString())
             RecordDbUtils.getInstance().save(record)
             EventBus.getDefault().post(MainThreadEvent(MainThreadEvent.RECORD_REFRESH_MY, record))
         } catch (e: Exception) {
@@ -317,18 +320,18 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
     }
 
     private fun showMusicDialog() {
-        musicDialog?.showBottomSheet()
+        musicDialog.showBottomSheet()
     }
 
     /**
      * 展示重新录制
      */
     private fun showRerecordDialog() {
-        AlertDialog.Builder(context)
+        OrangeDialog.Builder(context)
             .setTitle("重新录制")
-            .setMessage("确定删除当前的录音，并重新录制吗？")
-            .setNegativeButton("取消") { _, _ -> recordUtils!!.stopFullRecord() }
-            .setPositiveButton("重录") { _, _ ->
+            .setContent("确定删除当前的录音，并重新录制吗？")
+            .setNegativeText("取消").setOnNegativeCallBack { _, _ -> recordUtils!!.stopFullRecord() }
+            .setPositiveText("重录").setonPositiveCallBack {  _, _ ->
                 //可自行判断是否删除老的文件
                 oldRecord?.audioLength = 0
                 oldRecord?.audio_url = ""
@@ -345,11 +348,11 @@ open class MixRecordPage(private val context: AppCompatActivity, mRoot: ViewGrou
      */
     private fun showTipDialog() {
         onPause()//先暂停
-        AlertDialog.Builder(context)
+        OrangeDialog.Builder(context)
             .setTitle("温馨提示")
-            .setMessage("确定要停止录音吗？")
-            .setNegativeButton("停止录音") { _, _ -> recordUtils!!.stopFullRecord() }
-            .setPositiveButton("继续录音") { _, _ -> recordUtils!!.startOrPause() }
+            .setContent("确定要停止录音吗？")
+            .setNegativeText("停止录音") .setOnNegativeCallBack { _, _ -> recordUtils!!.stopFullRecord() }
+            .setPositiveText("继续录音") .setonPositiveCallBack { _, _ -> recordUtils!!.startOrPause() }
             .show()
     }
 
