@@ -1,21 +1,17 @@
 package me.shetj.mp3recorder.record.bean
 
-
-import io.reactivex.schedulers.Schedulers
-import me.shetj.base.s
+import io.reactivex.rxjava3.schedulers.Schedulers
+import me.shetj.base.S
 import me.shetj.base.tools.file.FileUtils
 import me.shetj.mp3recorder.record.bean.db.AppDatabase
 import me.shetj.mp3recorder.record.bean.db.RecordDao
-import me.shetj.simxutils.DbManager
-import me.shetj.simxutils.DbUtils
-import me.shetj.simxutils.ex.DbException
 import java.io.File
-import java.util.*
 
 class RecordDbUtils private constructor() {
     private val dbManager: RecordDao by lazy {
-        AppDatabase.getInstance(s.app).recordDao()
+        AppDatabase.getInstance(S.app).recordDao()
     }
+
     /**
      * 获取全部的录音
      */
@@ -29,46 +25,27 @@ class RecordDbUtils private constructor() {
 
 
     fun save(recordNew: Record) {
-        try {
-            dbManager.insertRecord(recordNew).subscribeOn(Schedulers.io()).subscribe()
-        } catch (e: DbException) {
-            e.printStackTrace()
-        }
-
+        dbManager.insertRecord(recordNew).subscribeOn(Schedulers.io()).subscribe()
     }
 
     /**
      * 更新录音
      */
     fun update(message: Record) {
-        try {
-            dbManager.insertRecord(message).subscribeOn(Schedulers.io()).subscribe()
-        } catch (e: DbException) {
-            e.printStackTrace()
-        }
-
+        dbManager.insertRecord(message).subscribeOn(Schedulers.io()).subscribe()
     }
 
     /**
      * 删除
      */
     fun del(record: Record) {
-        try {
-            dbManager.deleteRecord(record).subscribeOn(Schedulers.io()).subscribe()
-            FileUtils.deleteFile(File(record.audio_url!!))
-        } catch (e: DbException) {
-            e.printStackTrace()
-        }
+        dbManager.deleteRecord(record).subscribeOn(Schedulers.io()).subscribe()
+        FileUtils.deleteFile(File(record.audio_url!!))
 
     }
 
     fun clear() {
-        try {
-            dbManager.deleteAll().subscribeOn(Schedulers.io()).subscribe()
-        } catch (e: DbException) {
-            e.printStackTrace()
-        }
-
+        dbManager.deleteAll().subscribeOn(Schedulers.io()).subscribe()
     }
 
     companion object {
