@@ -11,14 +11,16 @@ import android.widget.LinearLayout
 import android.widget.SeekBar
 import android.widget.TextView
 import androidx.appcompat.widget.AppCompatImageView
-import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import me.shetj.base.tools.app.ArmsUtils
 import me.shetj.mp3recorder.R
-import me.shetj.mp3recorder.record.bean.Music
+import me.shetj.mp3recorder.record.bean.MusicQ
 import me.shetj.mp3recorder.record.utils.MixRecordUtils
 import me.shetj.mp3recorder.record.utils.Util
 import me.shetj.player.PlayerListener
 import me.shetj.recorder.mixRecorder.PlayBackMusic
+import me.shetj.recorder.core.OnVolumeChange
+import me.shetj.recorder.core.VolumeConfig
 
 /**
  * 背景音乐控制
@@ -42,9 +44,11 @@ class BackgroundMixMusicView @JvmOverloads constructor(context: Context,
     private var audioPlayer: PlayBackMusic?=null //播放器
     private var recordUtils: MixRecordUtils?=null //录音
     private var addMusicView: LinearLayout  ?=null//添加背景音乐的空间
-    private var music: Music?=null //背景文件相关
-    private var musicDialog: MusicListBottomSheetDialog?=null//选择背景音乐
-    private val max = 600f
+    private var music: MusicQ?=null //背景文件相关
+    private var musicDialog: MusicListBottomQSheetDialog?=null//选择背景音乐
+    private val max :Float by lazy { VolumeConfig.getInstance(context).getMaxVoice().toFloat() }
+    private val volumeConfig: VolumeConfig by lazy { VolumeConfig.getInstance(context) }
+
     init {
         //设置view
         val view = LayoutInflater.from(context).inflate(R.layout.bg_music_view, null)
@@ -59,6 +63,11 @@ class BackgroundMixMusicView @JvmOverloads constructor(context: Context,
         mIvChange.setOnClickListener(this)
     }
 
+    private val onVolumeChange: OnVolumeChange = object : (Float) -> Unit(){
+        override fun invoke(p1: Float) {
+            mSeekBar.progress = (p1*max).toInt()
+        }
+    }
 
     fun removeMusic() {
         if (visibility == View.VISIBLE){
@@ -89,32 +98,44 @@ class BackgroundMixMusicView @JvmOverloads constructor(context: Context,
         audioPlayer = recordUtils?.getBgPlayer()
         recordUtils?.setBackgroundPlayerListener(this)
         audioPlayer?.setLoop(true)
+        mSeekBar.max = volumeConfig.getMaxVoice()
+        mSeekBar.progress = volumeConfig.getCurVolume()
+        showVolumeString(mSeekBar, recordUtils)
         mSeekBar.setOnSeekBarChangeListener(object :SeekBar.OnSeekBarChangeListener{
             override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
-                mTvVoice.text = "$progress%"
-                recordUtils?.setVolume(progress/max)
+                seekBar?.let {
+                    showVolumeString(it, recordUtils)
+                }
             }
             override fun onStartTrackingTouch(seekBar: SeekBar?) {
             }
             override fun onStopTrackingTouch(seekBar: SeekBar?) {
                 seekBar?.let {
-                    recordUtils?.setVolume(it.progress/max)
-                    mTvVoice.text = "${it.progress}%"
+                    showVolumeString(it, recordUtils)
                 }
             }
         })
-        recordUtils?.setVolume(30/max)
     }
+
+    private fun showVolumeString(
+        it: SeekBar,
+        recordUtils: MixRecordUtils?
+    ) {
+        val fl = it.progress / max
+        recordUtils?.setVolume(fl)
+        mTvVoice.text = "${(fl * 100).toInt()}%"
+    }
+
     /**
      * 第2步设置背景音乐选择
      */
-    fun setDialog(musicDialog: MusicListBottomSheetDialog) {
+    fun setDialog(musicDialog: MusicListBottomQSheetDialog) {
         this.musicDialog = musicDialog
     }
     /**
      * isSave 是否保存，默认是保存
      */
-    fun setMusic(music: Music,isSave: Boolean = true){
+    fun setMusic(music: MusicQ, isSave: Boolean = true){
         if (music.url == null){
             //如果是错误数据，直接忽略
             return
@@ -127,6 +148,7 @@ class BackgroundMixMusicView @JvmOverloads constructor(context: Context,
         this.music= music
         mTvName.text = music.name
         mTvProgress.text = Util.formatSeconds3(0 / 1000) +"/"+  Util.formatSeconds3(music.duration.toInt())
+        mSeekBar.progress = volumeConfig.getCurVolume()
         music.startNoPlayMusic()
         if (visibility == View.GONE){
             TransitionManager.beginDelayedTransition(parent as ViewGroup?)
@@ -137,6 +159,15 @@ class BackgroundMixMusicView @JvmOverloads constructor(context: Context,
     }
 
 
+    override fun onAttachedToWindow() {
+        super.onAttachedToWindow()
+        VolumeConfig.getInstance(context).addChangeListener(onVolumeChange)
+    }
+
+    override fun onDetachedFromWindow() {
+        super.onDetachedFromWindow()
+        VolumeConfig.getInstance(context).removeChangeListener(onVolumeChange)
+    }
 
 
 
@@ -170,10 +201,10 @@ class BackgroundMixMusicView @JvmOverloads constructor(context: Context,
 
     }
 
-    private fun Music.startPlayMusic(): Unit? {
+    private fun MusicQ.startPlayMusic(): Unit? {
         audioPlayer?.let {
             if (!it.isPlayingMusic) {
-                audioPlayer?.setBackGroundUrl(url!!)
+                audioPlayer?.setBackGroundUrl(context,url!!,null)
                 audioPlayer?.startPlayBackMusic()
             }else{
                 if (it.isIsPause){
@@ -186,12 +217,12 @@ class BackgroundMixMusicView @JvmOverloads constructor(context: Context,
         return recordUtils?.setVolume(mSeekBar.progress / max)
     }
 
-    private fun Music.startNoPlayMusic(): Unit? {
-        audioPlayer?.setBackGroundUrl(url!!)
+    private fun MusicQ.startNoPlayMusic(): Unit? {
+        audioPlayer?.setBackGroundUrl(context,url!!,null)
         return recordUtils?.setVolume(mSeekBar.progress / max)
     }
 
-    override fun onStart(url: String, duration: Int) {
+    override fun onStart( duration: Int) {
         AndroidSchedulers.mainThread().scheduleDirect {
             mIvPlay.setImageResource(R.drawable.icon_pause_bg_music)
         }
@@ -226,9 +257,9 @@ class BackgroundMixMusicView @JvmOverloads constructor(context: Context,
 
     }
 
-    override fun onProgress(current: Int, size: Int) {
+    override fun onProgress(current: Int, duration: Int) {
         AndroidSchedulers.mainThread().scheduleDirect {
-            mTvProgress.text = Util.formatSeconds3(current / 1000) +"/"+ Util.formatSeconds3(size / 1000)
+            mTvProgress.text = Util.formatSeconds3(current / 1000) +"/"+ Util.formatSeconds3(duration / 1000)
         }
     }
 }
